Port 1.21.1 3eec0b8bc (partial): tooltip state consolidation + world target interface
- Add 3 world target key-mapping getters (getBookmarkWorldTarget, getShowWorldTargetRecipe,
  getShowWorldTargetUses) to IInternalKeyMappings.
- Add corresponding en_us + zh_cn translations for the new category and keys.
- BookmarkPermutationTooltipState: generalize sourceKey from int to Object (overload kept),
  matching the 1.21.1 API so callers can pass anything as the source identity.
- BookmarkCandidateTooltipHelper: unchanged call shape; uses the new state overload.

NOT ported (incompatible with 1.20.1 fork shape):
- getOrCreateTooltip helper move into the state (1.20.1 BookmarkIngredientKey has no
  typedIngredient() accessor — needs the broader BookmarkIngredientKey refactor).
- BookmarkConfig / BookmarkJsonConfig / BookmarkGroupConfigSerializer rewrite.
- WorldInputHandler and the world target keymapping registrations.
- BookmarkList / BookmarkGroup / BookmarkItemType refactor.
- NeoForge-only files (not present in 1.20.1).
- Test files.

+++ Gui/src/main/java/mezz/jei/gui/bookmarks/BookmarkCandidateTooltipHelper.java
@@ -64,7 +64,7 @@ public final class BookmarkCandidateTooltipHelper {
 			return;
 		}
 		int selectedIndex = resolvedKeys.indexOf(selectedKey);
-		int windowStart = tooltipState.updateStart((int) sourceKey, resolvedKeys, selectedIndex);
+		int windowStart = tooltipState.updateStart(sourceKey, resolvedKeys, selectedIndex);
 		tooltip.add(CandidateTooltipComponent.create(ingredientManager, candidates, selectedIndex, windowStart));
 	}
